feat(zkvms_host_io): Move metrics output from the benchmarkable macro to zkvsm_host_io lib
